# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export interface AdminCredentials {
  username: string;
  password: string;
}

export interface AdminToken {
  token: string;
  username: string;
}

export interface AdminUser {
  id: number;
  username: string;
  createdAt: string;
}

export interface Proxy {
  id: number;
  host: string;
  port: number;
  protocol: string;
  /** @nullable */
  username?: string | null;
  /** @nullable */
  password?: string | null;
  enabled: boolean;
  /** @nullable */
  label?: string | null;
  /** @nullable */
  lastUsedAt?: string | null;
  createdAt: string;
}

export type ProxyInputProtocol = typeof ProxyInputProtocol[keyof typeof ProxyInputProtocol];


export const ProxyInputProtocol = {
  http: 'http',
  https: 'https',
  socks5: 'socks5',
} as const;

export interface ProxyInput {
  /** @minLength 1 */
  host: string;
  /**
     * @minimum 1
     * @maximum 65535
     */
  port: number;
  protocol: ProxyInputProtocol;
  username?: string;
  password?: string;
  label?: string;
  enabled?: boolean;
}

export type ProxyUpdateProtocol = typeof ProxyUpdateProtocol[keyof typeof ProxyUpdateProtocol];


export const ProxyUpdateProtocol = {
  http: 'http',
  https: 'https',
  socks5: 'socks5',
} as const;

export interface ProxyUpdate {
  host?: string;
  port?: number;
  protocol?: ProxyUpdateProtocol;
  username?: string;
  password?: string;
  label?: string;
  enabled?: boolean;
}

export interface ProxyStats {
  total: number;
  enabled: number;
  disabled: number;
}

export interface ScraperUser {
  id: number;
  email: string;
  tokenBalance: number;
  /** @nullable */
  externalId?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface ScraperUserInput {
  email: string;
  tokenBalance?: number;
  externalId?: string;
  notes?: string;
}

export interface TokenAdjustment {
  /** Positive to add, negative to remove */
  amount: number;
  reason?: string;
}