# lib/api-zod/src/generated/types/proxy.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Proxy {
  id: number;
  host: string;
  port: number;
  protocol: string;
  /** @nullable */
  username?: string | null;
  /** @nullable */
  password?: string | null;
  enabled: boolean;
  /** @nullable */
  label?: string | null;
  /** @nullable */
  lastUsedAt?: Date | null;
  createdAt: Date;
}