# lib/api-zod/src/generated/types/proxyInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProxyInputProtocol } from './proxyInputProtocol';

export interface ProxyInput {
  /** @minLength 1 */
  host: string;
  /**
     * @minimum 1
     * @maximum 65535
     */
  port: number;
  protocol: ProxyInputProtocol;
  username?: string;
  password?: string;
  label?: string;
  enabled?: boolean;
}