# lib/api-zod/src/generated/types/proxyUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProxyUpdateProtocol } from './proxyUpdateProtocol';

export interface ProxyUpdate {
  host?: string;
  port?: number;
  protocol?: ProxyUpdateProtocol;
  username?: string;
  password?: string;
  label?: string;
  enabled?: boolean;
}