# lib/api-zod/src/generated/types/scraperUser.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ScraperUser {
  id: number;
  email: string;
  tokenBalance: number;
  /** @nullable */
  externalId?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: Date;
  updatedAt: Date;
}