# lib/api-zod/src/generated/types/scraperUserInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ScraperUserInput {
  email: string;
  tokenBalance?: number;
  externalId?: string;
  notes?: string;
}