# lib/api-zod/src/generated/types/tokenAdjustment.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TokenAdjustment {
  /** Positive to add, negative to remove */
  amount: number;
  reason?: string;
}